requirements.txt
django 
djangorestframework
psycopg2-binary
python-decouple
requests
beautifulsoup4
lxml
polls
peewee